Change command from "cc" to "club"
This is almost entirely just a personal preference, but given as there's
just the one namespace I think a two-character code where one of the
characters stands for "crossword" is insufficently unambiguous. I
acknowledge the commands are mostly set by vibes and the vibes are calling
me in this direction.