Make import of downloader modules much cleaner
Gets rid of a lot of the complicated introspection required to pick
up the available downloader classes, and also moves all of this
logic out of xword_dl.py.

This change could potentially be amended to add the plugins to the
downloader namespace as before, but after this change along with the
last several, this has no real benefit and leaves the namespace
cluttered.