CompilerDownloader: return binary response from solver URL (#278)
CrosswordCompiler (CC) games are (or contain) XML encoded puzzles.
In some cases, as with Daily Pop, these puzzles are sent with
headers that don't specify the character encoding. Nevertheless,
the XML declaration in the file specifies it correctly - e.g.

    encoding="UTF-8"

When this happens, because XML has the mimetype text/xml, requests
will (correctly according to the spec) assume the ISO-8859-1
encoding. If a different encoding is used for the file, this will
result in mojibake.

The decoding step is actually unncessary with CC puzzles because
they are passed to an XML parsing library (xmltodict, using expat)
which is capable of parsing the encoding from the declaration,
rather than relying on the HTTP header.

This commit removes the unnecessary decode step. Fixes #275.